Enable the user to set a payload for MQTT messages
As discussed in #25, the current payload was nearly constant. With this
patch, the user is able to set a constant payload. The PayloadGenerator
interface should allow us to use generators with probabilistic settings.
Like generating payloads according to a normal distribution.
Implementing such dynamic generators is left for a later PR.