feature: allow random size payload generation
by setting min-msg-size and max-msg-size activate a payload generator which will generate payloads with a random size between those two values

context: in our project we have a requirement that a broker must handle load where messages sizes are distributed between a min and max value. This change allows simulating that in a crude way.